lambda-metrics-demo/lambda/requirements.txt
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-exporter-otlp==1.16.0